Added a test for example mismatches as expectations printed on stdout
